Refactors auction controller
- Moves validation to middleware
  Unsure how to pass the data fetched to the controller.
- Extracts NE call to library
