+++ controllers/auctions.js
@@ -1,9 +1,9 @@
 const Joi = require("joi");
 const axios = require("axios");
 const Offer = require("../models/offer");
-const User = require("../models/user");
 const formatDistanceToNow = require("date-fns/formatDistanceToNow");
 const formatDate = require("date-fns/format");
+const { createAuction } = require("../lib/ne");
 
 const displayDate = (dateString) => {
   return formatDate(new Date(dateString), "yyyy-MM-dd HH:mm");
@@ -13,133 +13,17 @@ const NE_BASE_URL =
   process.env.NEGOTIATION_ENGINE_BASE_URL || "http://localhost:5000";
 
 /**
- * Validates a single offer auction. Checks that
- * - Offer exists
- * - The offer creator is the same as the auction creator.
- *
- * @param {string} creatorUsername
- * @param {string} offerId
- * @returns Offer from Digiprime and the auction type (Ascending, Descending)
- */
-const validateCreateFromSingleOffer = async (creatorUsername, offerId) => {
-  const offer = await Offer.findById(offerId).populate("author");
-  if (!offer) {
-    throw new Error("Offer not found");
-  }
-  const auctionType =
-    offer.costumer.toLowerCase() === "supply" ? "Ascending" : "Descending";
-
-  if (offer.author.username !== creatorUsername) {
-    throw new Error("Cannot create auction from other user's offer");
-  }
-
-  return { offer, auctionType };
-};
-
-/**
- * Validates members in a owned-offer auction. Checks that
- * - Creator is not part of members.
- * - All members exist in the Digiprime database.
- *
- * @param {string} creatorUsername
- * @param {string | string[]} members
- * @returns {string[]} member usernames
- */
-const validateMembers = async (creatorUsername, members) => {
-  let usernames = Array.isArray(members) ? members : [members];
-  usernames.forEach((username) => {
-    if (username === creatorUsername) {
-      throw new Error("Cannot add the creator as a member");
-    }
-  });
-
-  const users = await User.find({ username: { $in: usernames } }).exec();
-  if (users.length !== usernames.length) {
-    throw new Error("Could not find all passed members in user collection");
-  }
-  return usernames;
-};
-
-/**
- * Validates auction parameters for an auction containing multiple offers. Checks that
- * - All passed offer IDs exist as offers.
- * - All offers have the same reference sector.
- * - All offers have the same reference type.
- * - All offer creators are distinct.
- * - The auction creator is not a creator of any of the offers.
- *
- * @param {string} creatorUsername
- * @param {string[]} offerIds
- * @returns information about auction and the offers contained.
- */
-const validateCreateFromMultipleOffers = async (creatorUsername, offerIds) => {
-  // Get all the relevant offers to perform additional checks.
-  const offers = await Offer.find({ _id: { $in: offerIds } })
-    .populate("author")
-    .exec();
-
-  // Ensure all passed IDs are valid offers.
-  if (offers.length !== offerIds.length) {
-    throw new Error("Some or all Offer IDs are invalid");
-  }
-
-  // Perform additional checks.
-  // - Check all offers are of the same type.
-  // - Ensure all members are distinct.
-  // - Ensure auction creation is not a member.
-  let sector = offers[0].referenceSector;
-  let type = offers[0].referenceType;
-  let supplyOrDemand = offers[0].costumer;
-  let currentMembers = {};
-  for (let offer of offers) {
-    if (
-      offer.referenceSector !== sector ||
-      offer.referenceType !== type ||
-      offer.costumer !== supplyOrDemand ||
-      currentMembers[offer.author.username] ||
-      offer.author.username === creatorUsername
-    ) {
-      throw new Error("invalid");
-    }
-  }
-  const auctionType =
-    supplyOrDemand.toLowerCase() === "supply" ? "Descending" : "Ascending";
-
-  return { offers, sector, type, supplyOrDemand, auctionType };
-};
-
-const createQuerySchema = Joi.alternatives().try(
-  Joi.object({
-    from: Joi.string().valid("search").required(),
-    offerIds: Joi.array()
-      .min(2)
-      .items(Joi.string().pattern(/^[0-9a-fA-F]{24}$/, "offerId"))
-      .required(),
-  }),
-  Joi.object({
-    from: Joi.string().valid("offer").required(),
-    offerId: Joi.string()
-      .pattern(/^[0-9a-fA-F]{24}$/, "offerId")
-      .required(),
-  })
-);
-
-/**
  * Renders the page to create auctions.
  *
  * @param {*} req
  * @param {*} res
  */
 module.exports.create = async (req, res) => {
-  const q = await createQuerySchema.validateAsync(req.query);
-  const username = req.user.username;
-
-  if (q.from === "search") {
+  if (req.query.from === "search") {
     // Check if we should render the template for that contains multiple non-owned offers.
-    const { offers, sector, type, auctionType } =
-      await validateCreateFromMultipleOffers(username, q.offerIds);
+    const { offers, sector, type, auctionType } = req.body;
 
-    const combinedOfferIds = q.offerIds.join(",");
+    const combinedOfferIds = req.query.offerIds.join(",");
     res.render("auctions/create-multiple-offers", {
       offers,
       info: { sector, type, auctionType },
@@ -147,11 +31,7 @@ module.exports.create = async (req, res) => {
     });
   } else {
     // Otherwise it's from a single owned offer.
-    const { offer, auctionType } = await validateCreateFromSingleOffer(
-      username,
-      q.offerId
-    );
-    const users = await User.find({ username: { $nin: [username] } }).exec();
+    const { offer, auctionType, users } = req.body;
 
     res.render("auctions/create-single", {
       offer,
@@ -161,31 +41,6 @@ module.exports.create = async (req, res) => {
   }
 };
 
-const CreateAuctionSchema = Joi.alternatives().try(
-  Joi.object({
-    auctionTitle: Joi.string().required(),
-    closingTime: Joi.date().min(Date.now()).required(),
-    quantity: Joi.number().required(),
-    offerId: Joi.string()
-      .pattern(/^[0-9a-fA-F]{24}$/, "offerId")
-      .required(),
-    members: [
-      Joi.string().required(),
-      Joi.array().items(Joi.string()).required(),
-    ],
-    privacy: Joi.string().valid("Private", "Public").required(),
-  }),
-  Joi.object({
-    auctionTitle: Joi.string().required(),
-    closingTime: Joi.date().min(Date.now()).required(),
-    quantity: Joi.number().required(),
-    offerIds: Joi.array()
-      .min(2)
-      .items(Joi.string().pattern(/^[0-9a-fA-F]{24}$/, "offerId"))
-      .required(),
-  })
-);
-
 /**
  * Calls NegotationEngine to actually create the auction.
  *
@@ -193,19 +48,24 @@ const CreateAuctionSchema = Joi.alternatives().try(
  * @param {*} res
  */
 module.exports.createAuction = async (req, res) => {
-  const data = await CreateAuctionSchema.validateAsync(req.body);
   const username = req.user.username;
-  const fromSearch = data.members === undefined;
+  const { auctionTitle, quantity, closingTime } = req.body;
 
-  let params = {};
-  if (fromSearch) {
+  let data = {
+    room_name: auctionTitle,
+    quantity: quantity,
+    closing_time: new Date(closingTime).toISOString(),
+    templatetype: "article",
+  };
+
+  if (req.body.fromSearch) {
     // Auction created from an offer search.
-    const { offers, sector, type, auctionType } =
-      await validateCreateFromMultipleOffers(username, data.offerIds);
+    const { offerIds, offers, sector, type, auctionType } = req.body;
     const members = offers.map((offer) => offer.author.username);
 
-    params = {
-      articleno: data.offerIds.join(","),
+    data = {
+      ...data,
+      articleno: offerIds.join(","),
       reference_sector: sector,
       reference_type: type,
       auction_type: auctionType,
@@ -215,48 +75,33 @@ module.exports.createAuction = async (req, res) => {
   } else {
     // Auction created from a single owned offer.
     // Validate that all offers exists, and that all members exist.
-    const { offer, auctionType } = await validateCreateFromSingleOffer(
-      username,
-      data.offerId
-    );
-    let members = await validateMembers(username, data.members);
+    const { offer, auctionType, members, offerId, privacy } = req.body;
 
-    params = {
-      articleno: data.offerId,
+    data = {
+      ...data,
+      articleno: offerId,
       reference_sector: offer.referenceSector,
       reference_type: offer.referenceType,
       auction_type: auctionType,
       members: members,
-      privacy: data.privacy,
+      privacy: privacy,
     };
   }
 
-  // Create auction in NE
   try {
-    params = {
-      ...params,
-      room_name: data.auctionTitle,
-      quantity: data.quantity,
-      closing_time: new Date(data.closingTime).toISOString(),
-      templatetype: "article",
-    };
-
-    const urlParams = new URLSearchParams(params);
-    const response = await axios.post(`${NE_BASE_URL}/create-room`, urlParams, {
-      auth: { username },
-    });
-    // Response data contains a message, which contains the auction name and id.
-    // We are interested in the ID here.
-    // Example response: "The room auction #1 has been created id: 61e7f7e20daf6671113c4941"
-    const auctionId = response.data.message.split("id: ")[1];
+    const auctionId = await createAuction(username, data);
 
     req.flash("success", "Successfully created auction");
     res.redirect(`/auctions/${auctionId}`);
   } catch (error) {
-    console.error(error.response.data);
+    if (error.isAxiosError) {
+      console.error(error.response.data);
+    } else {
+      console.error(error);
+    }
 
     req.flash("error", "Failed to create auction");
-    res.render("auctions/create");
+    res.redirect("/auctions");
   }
 };
